+++ NEWS.md
@@ -1,8 +1,18 @@
-# mRchmadness 1.0.1
+# mRchmadness 1.0.2 - March 12, 2018
+
+## New data
+* bracket.women.2018
+* pred.538.women.2018
+
+## Shiny updates
+* App flipped to 2018
+* App is men's only in this release, pending additional women's data
+
+# mRchmadness 1.0.1 - March 11, 2018
 
 ## New data
 * bracket.men.2018
-* bracket.women.2018
+* bracket.women.2018 - from Charlie Creme's bracketology page on ESPN
 * games.men.2018
 * games.women.2018
 * pred.538.men.2018
@@ -10,4 +20,3 @@
 
 ## New features
 * handling of unplayed first-round games
-